feat: parallel WAL restore
When the `maxParallel` parameter is specified, the instance manager now
use a parallel wal-restore algorithm to prefetch WAL files after the one
requested by PostgreSQL.

Having done that, the instance manager will use the `maxParallel`
parameter specified in the `BarmanObjectStorage` configuration used for
WAL archiving.

Parallel WAL restore is not used when bootstrapping a new cluster,
as in that case, barman-cloud-restore is invoked directly.

Co-authored-by: Philippe Scorsolini <philippe.scorsolini@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>