feat: remove evicted Pods
When the Kubelet evicts one Pod because of node pressure, that pod
remains in the known resources with "Failed" phase and "Evicted" reason.
The operator will promote another instance, if needed, and then stop
working complaining that there is one Pod not active.

The only way an user could heal such a cluster is to manually delete the
evicted Pod.

With this patch the operator will do exactly that, after having put an
event in the cluster and have logged a warning message describing what's
happening.

Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Valerio Del Sarto <valerio.delsarto@enterprisedb.com>