feat: enhance `status` command of `cnp` plugin
Improve the `status` command of the `cnp` plugin for `kubectl`
by adding a new section with replication status (taken from
`pg_stat_replication`), as well as providing information about
the replication role of each Postgres instance in the cluster.

The patch also adds display of the total size of the databases
hosted in each instance.

Co-authored-by: Philippe Scorsolini <philippe.scorsolini@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>