fix: rolling deployment for replica clusters
When we switch over to a newly designated primary in a replica
cluster, we also set archive_mode to always in the
postgresql.auto.conf file. That will require a restart of the new
primary server, and that makes the switchover procedure unable to
correctly terminate since a promotion directly means a Pod requiring a
restart.

Besides that, when choosing the Pod to promote, we may promote the Pod
which already is the designated primary: we assume that the Pod list
always contains the primary Pod as the first entry, and this is not
always true for replica clusters, where every Pod is a replica and
replicas may be totally aligned.

We handle this case in the rolling update implementation with the
following patch and change the configuration strategy to set
`archive_mode` to `always` for any PostgreSQL instance in a replica
cluster. We then restrict the uploading of the WAL only from the
current and target designated primary. A WAL may be uploaded twice
during switchovers, which is not an issue.

Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>