feat: Set constant monitoring configmap name in cluster namespace
There were issues with updating the monitoring query configmap name in
the cluster CR when the configmap was updated in the operators env
variable. This was causing old configmaps to be left in the CR and new
ones to not be created when the operator restarted. Setting a consistent
name for the configmap in the namespace that the cluster is in fixes
this problem.

Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>