feat: cleanup default monitoring config maps after upgrade
Previously the operator, if configured with MONITORING_QUERIES_CONFIGMAP="A",
injected into every Cluster not explicitly disabling default monitoring queries
(`.spec.monitoring.disableDefaultQueries: true`) the configmap "A" into
the customQueriesConfigMap list and cloned the secret into the cluster namespace
with name "A".

With a previous patch we changed the target configmap name from being cloned with
the same name as the source configmap to being cloned always with the same target
name: `postgresql-operator-default-monitoring`.

With this patch, we introduce a cleanup function for the previously
created config maps which will take care of both deleting the old configmap
and removing it from the Cluster spec. This function should only be kept
to handle the upgrade from these versions and should be removed as soon
as it will be considered not required anymore.

Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>