feat: add backup destination sanity check for new clusters
Forgetting to remove the content of the backup destination when a cluster
is initialized from scratch is a common mistake. This patch prevents that
by using `barman-cloud-check-wal-archive` command to verify the emptiness
of the cloud destination just after the cluster initialization.

If the destination is not empty, we refuse to archive WALs in it.

For the moment, you can only see the reason for the failure in the logs.
A further patch will add some way to make this failure more explicit.

Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>