feat: try switching to streaming when wal-restore lookahead hits a file miss
This patch introduces a heuristic that allows PostgreSQL to switch from archive
to streaming replication mode in those clusters with a high WAL file production
rate where parallel archive and recovery are in use.

It does so by making the `wal-restore` subcommand fail in case any of the
requested WAL files (based on the `maxParallel` parameter) is not found in the
WAL archive (recovery object store). This implies that the replica is “almost”
up to date with the primary instance.
The failure of the `wal-restore` command implicitly suggests PostgreSQL to
attempt attaching itself again to the primary instance via streaming
replication, with all the resulting benefits.

Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>