chore: use controller-runtime in the instance manager
Refactor the instance manage code to replace the custom logic to watch the `Cluster`
resource with `controller-runtime`.

The primary reason for this patch is to avoid stale data reads from the API server
cache when a Pod starts.

As part of this change and the refactoring activity, we have introduced a clearer
and better separation of responsibilities.

This is just a first attempt, and more changes are needed in the future.

Co-authored-by: Philippe Scorsolini <philippe.scorsolini@enterprisedb.com>
Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>