fix: reverting to arbitrary order for `primary_conninfo`
The old code was totally correct and sound. However, changing the order
of the keys in `primary_conninfo` would require an unneeded restart
of clusters based on PostgreSQL 12 and older.

Besides that restart being unneeded, this would break the E2E tests
that expect that, after an online upgrade of the instance manager,
Pods continue to be up and running.

This doesn't happen from PostgreSQL 13 on, since from this version the
`primary_conninfo` string can be changed simply by reloading the
server.