fix: support update of Hot Standby sensitive parameters
Some PostgreSQL settings determine the size of the shared memory for tracking
internal information. For replicas to correctly work they must be greater or
equal on a standby than on the primary.

Given that, the CNP plain rollout strategy won't work when lowering those
values, as changes would be applied to replicas first and then, after a
switchover has been processed, to the old primary node.

The affected parameters are:

- `max_connections`
- `max_prepared_transactions`
- `max_locks_per_transaction`
- `max_wal_senders`
- `max_worker_processes`

This patch introduces a new feature in the instance manager, allowing it to
restart the PostgreSQL instances when requested. A new fencing feature has been
introduced too, mocking the result of the liveness probe while PostgreSQL is
restarting to avoid the Pod being restarted by the Kubelet.

With this commit, the previously discussed configuration changes will be
applied on the primary node and, after the primary node is restarted, the
change will be applied in replicas too. This fixes a well-known limitation
of Cloud Native Postgres - where the only workaround was to scale down to 1
instance, change the setting, and scale up.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>