chore: add Pyroscope support for setup-cluster.sh script
The cluster setup script now supports the option `-p` to deploy the cluster
with Pyroscope, useful to get CPU and memory usage from the operator.
Also deploy the operator with `debug` logs and pprof enabled, so that Pyroscope
can pull data from the pprof server enabled in the operator and get debug log
messages - if needed for debugging purposes.