feat: adopt open source license and change IP assignment
Adopt Apache License 2.0.

Transfer IP assignment from “EnterpriseDB Corporation” to “The
CloudNativePG Contributors”, the openly governed and vendor neutral
community that sparks from this commit (see the GOVERNANCE.md file).

As members and contributors of CloudNativePG, we are all extremely
grateful to EDB for open sourcing this Postgres operator for
Kubernetes, and for giving us the opportunity to apply for the CNCF
Sandbox.

Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>