chore: add directory content after TARGETARCH in GoReleaser
The new version of GoReleaser started to add `_v1`, `_v2` or `_v3`
for the new support of GOAMD64 added in 1.18. This patch aims to
add anything after the TARGETARCH variable en before the `/` in
the build directory created by GoReleaser. This change assumes we
are not supporting more than one flavor per architecture.

Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>