fix: enable recovery with just `targetTLI` option
With this patch, we handle correctly `targetTLI` also when it is the only
parameter specified as a recovery target. We have also removed support for
`current`, so that only `latest` or integers will be accepted (previously it
was handled as a synonym for `latest` and could lead to unexpected behaviours).

Co-authored-by: Hai He <hai.he@enterprisedb.com>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Co-authored-by: Valerio Del Sarto <valerio.delsarto@enterprisedb.com>