feat: fencing limitation removed
With this patch we push down the fencing condition, so that fencing a
single instance will not anymore imply the whole cluster being fenced.
This way if a follower instace is fenced, in case of falure of the
primary instance, failovers will still be performed. In case of
deletion of a fenced instance, the pod will be recreated, relabelled
according to the previous role and if it was a primary instance, the
operator will not perform a failover as before.

Signed-off-by: Philippe Scorsolini <philippe.scorsolini@enterprisedb.com>