fix: requeue renconciliation loop if the RBAC for backups is not yet created (#106)
During an instance manager reconciliation loop, the RBAC for the
secrets needed by backup might not be ready yet. In this case, the
retrieval of the secret will result in a forbidden error that will
then be ignored by the process, leaving the instance manager cache in
an inconsistent state.

To avoid being stuck in an inconsistent state, we queue a new
reconciliation loop in 30 seconds if we get a permission error.

This patch also includes a better organization of the cache
reconciliation logic.

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Signed-off-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Jaime Silvela <jaime.silvela@enterprisedb.com>