fix: ensures that operator pods always have the latest certificates (#74)
This patch mounts the secret containing the certificate in the operator
deployment and marks it as optional to allow the first boot. The kubelet
will now make sure to update the mounted files each time the secrets get
updated.

This patch also reorganizes the certificate package by unexporting some
methods and by improving its logic encapsulation.

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Tao <tao.li@enterprisedb.com>