feat: offline data import and major upgrades for Postgres (#224)
Provide a declarative way to import existing Postgres databases in
CloudNativePG using offline migrations - which also cover major
upgrades of Postgres databases.

IMPORTANT: Offline means that, with this method, applications must
stop their write operations to the source until the database is
migrated to the new cluster. Offline migrations are in contrast to
online migrations with 0 cutover time - which are not available yet in
CloudNativePG.

Offline migrations extend the `initdb` bootstrap method of
CloudNativePG to enable users to create a new PostgreSQL cluster using
a logical snapshot of the data available in another PostgreSQL
database - which can be accessed via the network through a superuser
connection.

Import is from any supported version of Postgres and relies on
`pg_dump` and `pg_restore` to be executed from the new cluster primary
for all databases part of the operation and, if requested, for roles.

Two import types are introduced:

- *microservice*: select the single database to import into the new
   cluster as the main application database
- *monolith*: select one or more databases to import (including “*”
   wildcard), as well as roles (superuser privileges are removed)

The microservice approach is the preferred method to import databases
in CloudNativePG. In contrast, the monolith approach is more aligned
with a shared PostgreSQL instance having multiple databases that DBAs
are familiar with. More information is in the documentation included
in this commit.

Closes #9

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Tao Li <tao.li@enterprisedb.com>
Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Signed-off-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Signed-off-by: Jitendra Wadle <jitendra.wadle@enterprisedb.com>
Signed-off-by: Hai He <hai.he@enterprisedb.com>
Signed-off-by: John Long <john.long@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Signed-off-by: Danish Khan <danish.khan@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Tao Li <tao.li@enterprisedb.com>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Co-authored-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Co-authored-by: Jitendra Wadle <jitendra.wadle@enterprisedb.com>
Co-authored-by: Hai He <hai.he@enterprisedb.com>
Co-authored-by: John Long <john.long@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Danish Khan <danish.khan@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>