fix: version upgrade logic in patch #198 (#330)
The part of pull request #198 regarding the upgrade from a version
setting reasons containing space is wrong because it is impossible to
update the status in a defaulting webhook or with a simple Update
call. Moreover, the setDefault function misses the difference because
it only compares the content of the spec section.

In most cases, the previous code works anyway because the Cluster
reconciliation loop must update the status after an operator
upgrade. However, in rare cases, it could lead to a cluster that is
impossible to upgrade.

This patch fixes the issue by moving the upgrade code from the webhook
to the main controller reconciliation loop.

Closes #329

Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>