fix: use backup ID in Point in Time Recovery (#270)
Until now, we were not properly supporting point-in-time recovery when
executed specifying a transaction ID, a named restore point, or an
immediate target. The reason is that we would use the last available
backup in the catalog in this case. Effectively, only `targetTime`
(most popular) and `targetLSN` were working.

The fix introduces a new parameter called `backupID` in the
`recoveryTarget` section.

The `backupID` parameter is empty by default, meaning that the
operator automatically detects the backup from which to start the
recovery in case `targetTime` or `targetLSN` are
specified. Alternatively, the operator uses the latest available
backup as a fallback.

If `backupID` is present, the operator will always use that backup ID
as the start of the recovery process, leaving the end-user responsible
for correctly picking the base backup for recovery.

**IMPORTANT:** Although this is normally seen as a new feature, we
have considered this a bug fix and made it eligible for backporting
(the problem exists in 1.15.x and must be solved).

Closes #233

Signed-off-by: Jonathan Battiato <jonathan.battiato@enterprisedb.com>
Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Signed-off-by: Philippe Scorsolini <p.scorsolini@gmail.com>
Signed-off-by: John Long <john.long@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>

Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Co-authored-by: Philippe Scorsolini <p.scorsolini@gmail.com>
Co-authored-by: John Long <john.long@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>