fix: support application database monitoring for replica clusters
This patch enables CNPG to execute monitoring queries against the
application database for replica clusters, bootstrapped with the
`pgbasebackup` and the `recovery` bootstrap method.

We are fixing the behavior of `GetApplicationDatabaseName` too, directly
getting the database name (and owner) regardless of whenever the
application database need to be created or not.

Closes #345

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>