feat: add ready-to-test command and independent container registry
We introduce a new behaviour in this commit:

1.- The commits now can run in the forked repo, using the external
    user registry to push the image and the run the E2E.
2.- A new command `ready-to-test` that use the `pull_request_target`
    event to run the test and run the E2E test on the PR, this will
    allow forked repo to run test from the PR when a user allows it.

Peevioulsy we were not avaible to run the E2E on PR from forked repo
due to the fixed use of the cloudnative-pg repo and since the
`pull_request` event doesn't provide a GitHub token with write permissions
for packages.

We still allow running E2E test from internal branch only when the `test`
command is used, since it will trigger the event dispatch with the client
payload and is used to verify that comes from the internal repo.

Closes #426

Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>