Update connection_pooling.md
Hi! This is my first attempt at contributing to an open source project, so I apologize for any mistakes I make.

I looked at the source code and i noticed that the problem occurs only when the code blocks it's inside a list. 
I shifted left the triple backticks and not touched the text inside so that the code blocks starts from the beginning of the page but the text inside remain well formatted.

Signed-off-by: Enrico Lumini <54529567+Enry58@users.noreply.github.com>