Single replica breaking OCP upgrade
When running a single replica  POC the PDB will not allow eviction. This means OCP cannot upgrade the cluster unless the it is in a maintenance window. The problem with this logic is OCP administration and Postgres administration may not be in sync.  The proposal is to separate the single replica logic  from the maintenance window logic. 

Signed-off-by: Aaron M. Cohen <aaronmcohen@gmail.com>