feat: upgrade Kubebuilder version
Upgrade of Kubebuilder to version 3.5.0 along with a couple of other
tools that need to be upgraded.

This upgrade will allow to work with multi-group APIs in the future
if we need it and also to new comers that know kubebuilder 3.x to
join to the project in the future.

Closes #482

Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>