chore: add basic script to report diffs release branch <-> main
This helper script must be run inside a release branch and reports
in Markdown format the pull requests the are not present in both
branches. This hopefully will help maintainers keep track of which
commits to backport, while waiting for an automated mechanism
that, once a commit is merged in main, issues a PR on each requested
release.

Closes #552

Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>