chore: rename managedResource `pods` into `instances`
This patch assigns a more appropriate name to the pods contained inside
managedResources. The operator is creating roles for instances, and it's
creating Jobs to initialize the PVCs, and those jobs are creating Pods.

The main target of this patch is avoiding confusing the Pods created by
the operator for the instances with the ones created by the Jobs too.

The renaming was also applied to functions name that were meant to
operate only on instances.

Closes #550

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: John Long <john.long@enterprisedb.com>