chore: adds `cnpg.io/podRole` label
This patch adds the label `cnpg.io/podRole` that can be used when
creating a pod resource. This allows us to properly identify a Pod since
its creation without relying on labels added during the reconciliation
cycle.

Before this patch it was not possible to identify an instance pod before
it received the `role` label from a reconciliation cycle.

Closes #550

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>