fix: set explicit permissions on manager binary
This patch explicitly sets the permissions of the manager binary copied
in the image, in order to be sure the user used then has the required
exec privileges to actually run it at startup.
Moreover, the distroless/static:nonroot image provides a
nonroot:nonroot user, that can be used instead of the usual 1001 user.

Signed-off-by: Philippe Scorsolini <p.scorsolini@gmail.com>