feat(plugin): Add `hibernate on` to `cnpg` plugin
Introduce cluster hibernation which destroys all the resources generated by the
cluster, except the PVCs that belong to the PostgreSQL primary instance.

You can hibernate a cluster with:

```
kubectl cnpg hibernate on <cluster-name>
```

This will:

1. shutdown every PostgreSQL instance
2. detach the PVCs containing the data of the primary instance, and annotate
   them with the latest database status and the latest cluster configuration
3. delete the `Cluster` resource, including every generated resource - except
   the aforementioned PVCs

When hibernated, a CloudNativePG cluster is represented by just a group of
PVCs, in which the one containing the `PGDATA` is annotated with the latest
available status, including content from `pg_controldata`.

A cluster having fenced instances cannot be hibernated, as fencing is
part of the hibernation procedure too.

A future commit will implement the `hibernate off` command to resume the
cluster.

Closes #781

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>