fix: avoid failovers when Pod readiness status is stale (#890)
The operator relied on the Pod's readiness status to determine whether
the current primary was healthy. Still, sometimes it will be different
from the actual PostgreSQL status since it's checked every
`periodSeconds` by the Kubelet.

For example, it can happen when lifting fencing from the primary node;
The Kubelet still hasn't marked the Pod as ready, leading to an
immediate failover after the event.

With this patch, we exclusively use the status collected from the
instance manager to elect the new primary. If the Pod is not ready
from the point of view of Kubernetes, we'll wait until Kubernetes
refreshes its view.

Closes: #889

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>