chore: add HOME variables for MinIO and Azurite client
Currently, both storage client were not setting a default HOME path for
the user, this means that they were using `/` as the HOME path, this
can be considered as a bad practices.

This patch creates an emptyDir volume and mounts it with the name of `mc` and
`azurite` in `/` and set the HOME environment variable to that path.

Closes #905

Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>