fix: automatic PostgreSQL version update regexp
The regexp replacing the previous-latest version with the new one was
updating not related versions, like the CNPG version 1.15.0 to 1.15.1

The regexp is now more precise and only updates places where a space
or a colon character precedes the Postgres version.

We also exclude all the versions in the release notes.

Closes #997

Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>