feat: improve compatibility with Istio (#967)
We terminate the Istio proxy-agent via the `/quitquitquit` endpoint before the
initdb, join, pgbasebackup, restore processes terminate.

This behavior is needed for the corresponding Job to be marked successful,
as discussed in the [relative Istio issue](https://github.com/istio/istio/issues/11045).

We ensure the API server is reachable too before starting the processes,
allowing istio-proxy to startup correctly.

Closes: #801

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Hai He <hai.he@enterprisedb.com>
Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>

Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>