fix: use correct size to update WAL PVC
Previously we were using the size of the main PVC to update the one
dedicated to WAL storage. This led to "-wal" PVCs being created with
the correct size at first, but then, in case the main PVC had a
different size, to be resized to the main PVC's size.

Signed-off-by: Philippe Scorsolini <p.scorsolini@gmail.com>