chore: add support for environment variables
Now we add two new elements that we pass to the cluster pods: `Env` and
`EnvFrom` both of them have the same behavior as the one referenced for the
Pods in the Kubernetes API for `EnvVar` and `EnvFrom`.

The field `EnvFrom` will be directly copied to the Pod spec.

In our case, `Env` will not overwrite the enviorment variables needed
by the operator to manage PostgreSQL properly, it will discard any variable
that we already use internally.

Closes: #1090

Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>