refactor: add pkg/reconciler/persistentvolumeclaim (#1288)
This patch moves in a separate package the code that creates and
reconciles the PersistentVolumeClaims resources.

The main goal of this is to improve the readability and the maintainability
of the code, acting as a substrate to be used by the separate WAL storage
management features.

The same logic is now used for the PGDATA and the WAL volumes, resulting
in a coherent reconciliation and fixing some bugs that were affecting the
WAL volumes.

Closes:  #1287

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
(cherry picked from commit 42be16aae9fdadd62bf95a63e36fac6005986cba)