test(e2e): enhance feature_type to be able to ignore upgrade test (#1406)
This will add the possibility to ignore the upgrade test when running the `/test` command
like the the follow `/test feature_type=!upgrade` to avoid the upgrade test.

Also, when running `test feature_type=smoke` it will only run the smoke test and will
not try to run the upgrade test.

Signed-off-by: Tao Li <tao.li@enterprisedb.com>