ci: improve the backport process (#1436)
Previously we were not backporting every PR that started with the word "feat:",
this was causing a mess since some features needed to be backported.

Now, by default, we backport every PR and it's up to the reviewers to set
the proper labels for the backport or to decide to not backport at all.

Closes #1435

Signed-off-by: Tao Li <tao.li@enterprisedb.com>
(cherry picked from commit 44d194afde25265dfc199e708fb66f71225d723d)