refactor: centralize PVCs management (#1340)
Refactor the existing imperative code that is responsible for managing
PersistentVolumeClaim resources to support reconciliation and, at the
same time, improve source maintainability and reduce the cognitive load.

We have shifted the logic from an “instance pod” centric one to a
“PVC group” centric one, ensuring that from an existing group of
PVCs we can start up an instance pod - provided PostgreSQL data
is consistent.

The code is now entirely in a separate reconciler package.

This commit is the foundation for future improvements involving PVC
management, including enhanced control of initialization/bootstrap
procedures and support for tablespaces.

Closes #1321

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>