feat: Introduce failover delay during OnlineUpgrade phase
Previously, the online upgrade process could trigger failover logic unnecessarily due to delays in the spin-up time of the instance manager. To prevent this issue, a failover delay has been added specifically for the OnlineUpgrade phase. This new feature ensures that failover logic is only triggered when necessary, resulting in a more efficient upgrade process.

Closes #1727

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>