feat: kubectl-cnpg psql subcommand (#1668)
With `kubectl cnpg psql` the user can start a `psql` session connected to a
CloudNativePG cluster, by default, to the primary pod of this cluster.

The user can also use `--replica` to connect to the first replica pod in the
pod list of the cluster. This pod can't be selected for the moment.

Closes: #1667

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@gmail.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>