fix: use backup name to match resources in the backup object store (#1650)
Barman 3.3 introduced support for the `--name` option in
`barman-cloud-backup` to proactively set the name of a
base backup and reference it when needed (e.g. restore).

If the operand contains such version of Barman, CloudNativePG
determines the name of the backup before executing it,
changing the previous behavior in which Barman was
generating the ID of the backup.

This commit removes the need for the previous heuristics that
was associating the last `Backup` resource to the last backup
in the object store, potentially leading to false matches in
the rare case of concurrent backups.

When available, the status of the backup resource now contains
its `backupName`.

Closes #1523

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Signed-off-by: John Long <john.long@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: John Long <john.long@enterprisedb.com>
(cherry picked from commit c59cad6c3982618657282163e8a4bb7833948874)