fix: package name for .tar.gz files (#1767)
After upgrading to GoReleaser 1.15.0 we removed the `_version_` in the
package name for the tarball files. This broke the install script and
also the Arkade app that was relying on this name to download the archive.

Closes #1766

Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
(cherry picked from commit 6b041df3b95b5f3b7c1850af874c88bfe0574088)