feat: declarative role management (#1524)
Introduce the `managed.roles` stanza in the `Cluster` spec to provide
full lifecycle management of database roles, by providing an abstraction
to the related DDL commands, such as `CREATE ROLE` and `ALTER ROLE`,
with the current limitation of the `IN ROLE` relationships (which will be
included in a future patch).

This feature enables declarative management of existing roles, as well as the
creation of new roles if they are not already present in the database. Role
creation will occur after the database bootstrapping is complete.

Here is an excerpt that briefly highlights this feature:

```yaml
apiVersion: postgresql.cnpg.io/v1
kind: Cluster
spec:
  # <snip>
  managed:
    roles:
    - name: dante
      ensure: present
      comment: Dante Alighieri
      login: true
      superuser: false
```

Declarative role management ensures that PostgreSQL instances align with the
spec. If a user modifies role attributes directly in the database for a role that
is in the `managed` stanza, the CloudNativePG operator will revert those
changes during the next reconciliation cycle.

The CRD status includes a section for the managed roles' status, as shown below:

```yaml
  roleStatus:
    not-managed:
    - app
    pending-reconciliation:
    - dante
    reserved:
    - postgres
    - streaming_replica
```

In terms of backward compatibility, declarative role management is designed
to ignore roles that exist in the database but are not included in the spec.
The lifecycle of these roles will continue to be managed within PostgreSQL.

This commit is the first important step towards declarative management of
global objects in PostgreSQL, namely roles, databases and tablespaces, through
CloudNativePG.

Closes #1515

Signed-off-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Signed-off-by: Tao Li <tao.li@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Tao Li <tao.li@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>