feat: use separate transactions to reconcile users's credentials
The operator codebase now uses separate transactions to reconcile the
credentials of multiple users.

Before this patch, the operator used a single transaction for all roles.
A failure in the synchronization of a role led to a failure in the
synchronization of every role, being the transaction rolled back as a
whole.

The patch also includes a few fixes in the unit tests of related
functions.

See: #1881

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>