feat: use separate transactions to reconcile users credentials (#2004)
The operator codebase now uses separate transactions to reconcile the
credentials of multiple users.

Before this patch, the operator used a single transaction for all roles.
A failure in the synchronization of a role led to a failure in the
synchronization of every role, being the transaction rolled back as a
whole.

The patch also includes a few fixes in the unit tests of related
functions.

See: #1881

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Tao Li <tao.li@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>

Co-authored-by: Tao Li <tao.li@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
(cherry picked from commit 3571e4eb6ec383fec5eed44b520894df64c107be)