Prevent creating Kubernetes client when calling completion command (#1937)
When the `completion` command was called while it was used inside the
`~/.basrc` file or a `~/.profile` this was useless since we were not
able to output the completion becase we were trying to create a Kubernetes
client all the time, which it's not the case during the load of a profile
configuration, since this could happen even before there's any Kubernetes
cluster configured in place.

The previous example also applies to the `version` command.

Closes #1932

Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
(cherry picked from commit 842259fbd66998ef1e75c5bc3b2fdfd217a0b6ec)